Struct core_traits::ValueTypeSet

pub struct ValueTypeSet(pub EnumSet<ValueType>);

Methods

impl ValueTypeSet

pub fn any() -> ValueTypeSet

pub fn none() -> ValueTypeSet

pub fn of_one(t: ValueType) -> ValueTypeSet

Return a set containing only t.

pub fn of_numeric_types() -> ValueTypeSet

Return a set containing Double and Long.

pub fn of_numeric_and_instant_types() -> ValueTypeSet

Return a set containing Double, Long, and Instant.

pub fn of_keywords() -> ValueTypeSet

Return a set containing Ref and Keyword.

pub fn of_longs() -> ValueTypeSet

Return a set containing Ref and Long.

impl ValueTypeSet

pub fn insert(&mut self, vt: ValueType) -> bool

pub fn len(&self) -> usize

pub fn union(&self, other: &ValueTypeSet) -> ValueTypeSet

Returns a set containing all the types in this set and other.

pub fn intersection(&self, other: &ValueTypeSet) -> ValueTypeSet

pub fn difference(&self, other: &ValueTypeSet) -> ValueTypeSet

Returns the set difference between self and other, which is the set of items in self that are not in other.

pub fn exemplar(&self) -> Option<ValueType>

Return an arbitrary type that's part of this set. For a set containing a single type, this will be that type.

pub fn is_subset(&self, other: &ValueTypeSet) -> bool

pub fn is_disjoint(&self, other: &ValueTypeSet) -> bool

Returns true if self and other contain no items in common.

pub fn contains(&self, vt: ValueType) -> bool

pub fn is_empty(&self) -> bool

pub fn is_unit(&self) -> bool

Important traits for Iter<E>
impl<E> Iterator for Iter<E> where
E: CLike, type Item = E;

pub fn iter(&self) -> Iter<ValueType>

impl ValueTypeSet

pub fn is_only_numeric(&self) -> bool